+++ src/commands/limits.ts
@@ -12,13 +12,17 @@ export const limits: Command = {
 	name: "limits",
 	description: "Display the song queue's submission limits.",
 	requiresGuild: true,
-	async execute({ guild, guildLocale, reply }) {
+	async execute({ type, guild, guildLocale, userLocale, reply }) {
 		const queueChannel = await getQueueChannel(guild);
 		if (!queueChannel) {
 			return await reply(t("common.queue.not-set-up", guildLocale));
 		}
 
 		const config = await getQueueConfig(queueChannel);
+
+		// We call `reply` with `ephemeral: true`, which has different behavior between
+		// slash-command interactions and message commands.
+		const locale = type === "interaction" ? userLocale : guildLocale;
 
 		// Read out the existing limits
 		const embed = new EmbedBuilder() //
@@ -30,28 +34,28 @@ export const limits: Command = {
 			switch (key.value) {
 				case "cooldown":
 					if (config.cooldownSeconds !== null && config.cooldownSeconds > 0) {
-						value = durationString(config.cooldownSeconds);
+						value = durationString(locale, config.cooldownSeconds);
 					} else {
 						value = "none";
 					}
 					break;
 				case "entry-duration":
 					if (config.entryDurationSeconds !== null && config.entryDurationSeconds > 0) {
-						value = durationString(config.entryDurationSeconds);
+						value = durationString(locale, config.entryDurationSeconds);
 					} else {
 						value = "infinite";
 					}
 					break;
 				case "entry-duration-min":
 					if (config.entryDurationMinSeconds !== null && config.entryDurationMinSeconds > 0) {
-						value = durationString(config.entryDurationMinSeconds);
+						value = durationString(locale, config.entryDurationMinSeconds);
 					} else {
 						value = "0";
 					}
 					break;
 				case "queue-duration":
 					if (config.queueDurationSeconds !== null && config.queueDurationSeconds > 0) {
-						value = durationString(config.queueDurationSeconds);
+						value = durationString(locale, config.queueDurationSeconds);
 					} else {
 						value = "infinite";
 					}