+++ src/commands/stats.ts
@@ -47,9 +47,9 @@ export const stats: Command = {
 		]);
 
 		// Average song length
-		const durationMsg = createPartialString(durationString(avgDuration));
+		const durationMsg = createPartialString(durationString(userLocale, avgDuration));
 		if (config.entryDurationSeconds !== null && config.entryDurationSeconds > 0) {
-			push(` (limit ${durationString(config.entryDurationSeconds)})`, durationMsg);
+			push(` (limit ${durationString(userLocale, config.entryDurationSeconds)})`, durationMsg);
 		}
 		embed.addFields({ name: "Average Length of Your Submissions", value: composed(durationMsg) });
 
@@ -77,7 +77,7 @@ export const stats: Command = {
 				timeSinceLatest !== null
 					? Math.max(0, config.cooldownSeconds - timeSinceLatest)
 					: 0;
-			const value = durationString(timeToWait);
+			const value = durationString(userLocale, timeToWait);
 			embed.addFields({ name: "Time Remaining on Cooldown", value });
 
 			// TODO: ETA to user's next submission would be nice here