+++ src/handleInteraction.ts
@@ -61,6 +61,8 @@ export async function handleInteraction(
 			channel = interaction.channel;
 		}
 
+		// TODO: Let the user specify a userspace locale override outside of their Discord client preference. `/prefs` command maybe?
+
 		const context: CommandContext = {
 			type: "interaction",
 			createdTimestamp: interaction.createdTimestamp,